chore(meta-service): remove obsolete snapshot v002 support
