+++ src/meta/types/src/lib.rs
@@ -27,7 +27,6 @@ mod grpc_helper;
 mod log_entry;
 mod message;
 mod operation;
-mod raft_snapshot_data;
 mod raft_txid;
 mod seq_num;
 mod time;
@@ -121,4 +120,3 @@ pub use crate::cmd::CmdContext;
 pub use crate::cmd::MetaSpec;
 pub use crate::cmd::UpsertKV;
 pub use crate::grpc_helper::GrpcHelper;
-pub use crate::raft_snapshot_data::SnapshotData;
